Внесла исправления по комментариям
